requirements.txt
altgraph==0.17.4
anyio==4.8.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.2
dill==0.3.9
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
keyboard==0.13.5
lxml==5.3.1
multiprocess==0.70.17
nest-asyncio==1.6.0
# Keep NumPy <2.0 for compatibility with paddlepaddle/paddleocr stack on Windows
numpy==1.26.4
# PaddleOCR 2.7.2 requires opencv-python <= 4.6.0.66
opencv-python==4.6.0.66
ordered-set==4.1.0
packaging==24.2
pathos==0.3.3
pefile==2023.2.7
pillow==11.1.0
pox==0.3.5
ppft==1.7.6.9
pycparser==2.22
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
pywin32==308
requests==2.32.3
scipy==1.13.1
scikit-learn==1.5.2
sniffio==1.3.1
tqdm==4.67.1
translators==5.9.9
typing_extensions==4.12.2
urllib3==2.3.0
winrt-runtime==2.3.0
winrt-Windows.Foundation==2.3.0
winrt-Windows.Foundation.Collections==2.3.0
winrt-Windows.Globalization==2.3.0
winrt-Windows.Graphics.Imaging==2.3.0
winrt-Windows.Media.Ocr==2.3.0
winrt-Windows.Storage==2.3.0
winrt-Windows.Storage.Streams==2.3.0
zstandard==0.23.0
mss==9.0.1
paddlepaddle==2.6.1
paddleocr==2.7.2
